requirements_test.txt
pytest==2.9.2
pytest-cov==2.3.0
pytest-django==2.9.1
pep8==1.7.0

